# Add benchmark for OGC API Features to measure + improve performance