--- a/engine/openapi.go
+++ b/engine/openapi.go
@@ -206,7 +206,7 @@ func (o *OpenAPI) ValidateRequest(r *http.Request) error {
 	return nil
 }
 
-func (o *OpenAPI) validateResponse(contentType string, body []byte, r *http.Request) error {
+func (o *OpenAPI) ValidateResponse(contentType string, body []byte, r *http.Request) error {
 	requestValidationInput, _ := o.getRequestValidationInput(r)
 	if requestValidationInput != nil {
 		responseHeaders := http.Header{HeaderContentType: []string{contentType}}