--- a/go.mod
+++ b/go.mod
@@ -5,8 +5,9 @@ go 1.21
 require (
 	dario.cat/mergo v1.0.0
 	github.com/BurntSushi/toml v1.3.2
-	github.com/PDOK/go-cloud-sqlite-vfs v0.2.4
+	github.com/PDOK/go-cloud-sqlite-vfs v0.2.5
 	github.com/creasty/defaults v1.7.0
+	github.com/docker/go-units v0.5.0
 	github.com/elnormous/contenttype v1.0.4
 	github.com/getkin/kin-openapi v0.122.0
 	github.com/go-chi/chi/v5 v5.0.10