--- a/ogc/features/datasources/geopackage/geopackage.go
+++ b/ogc/features/datasources/geopackage/geopackage.go
@@ -8,6 +8,7 @@ import (
 	"maps"
 	"os"
 	"path"
+	"sync"
 	"time"
 
 	"github.com/PDOK/gokoala/engine"
@@ -29,14 +30,18 @@ const (
 	bboxSizeBig      = 10000
 )
 
+var once sync.Once
+
 // Load sqlite (with extensions) once.
 //
 // Extensions are by default expected in /usr/lib. For spatialite you can
 // alternatively/optionally set SPATIALITE_LIBRARY_PATH.
-func init() {
-	spatialite := path.Join(os.Getenv("SPATIALITE_LIBRARY_PATH"), "mod_spatialite")
-	driver := &sqlite3.SQLiteDriver{Extensions: []string{spatialite}}
-	sql.Register(sqliteDriverName, sqlhooks.Wrap(driver, datasources.NewSQLLogFromEnv()))
+func loadDriver() {
+	once.Do(func() {
+		spatialite := path.Join(os.Getenv("SPATIALITE_LIBRARY_PATH"), "mod_spatialite")
+		driver := &sqlite3.SQLiteDriver{Extensions: []string{spatialite}}
+		sql.Register(sqliteDriverName, sqlhooks.Wrap(driver, datasources.NewSQLLogFromEnv()))
+	})
 }
 
 type geoPackageBackend interface {
@@ -75,6 +80,8 @@ type GeoPackage struct {
 }
 
 func NewGeoPackage(collections engine.GeoSpatialCollections, gpkgConfig engine.GeoPackage) *GeoPackage {
+	loadDriver()
+
 	g := &GeoPackage{}
 	g.preparedStmtCache = NewCache()