Benchmarking + performance tuning

```
pkg: github.com/PDOK/gokoala/ogc/features
              │ bench_run1.txt │           bench_run6.txt            │
              │     sec/op     │   sec/op     vs base                │
Features/1-16      57.94m ± 2%   31.40m ± 1%  -45.82% (p=0.000 n=10)
Features/2-16      68.86m ± 2%   41.24m ± 3%  -40.11% (p=0.000 n=10)
Features/3-16      59.59m ± 3%   32.00m ± 1%  -46.30% (p=0.000 n=10)
Features/4-16      68.86m ± 1%   41.82m ± 3%  -39.27% (p=0.000 n=10)
geomean            63.61m        36.28m       -42.96%

              │ bench_run1.txt │             bench_run6.txt             │
              │      B/op      │      B/op       vs base                │
Features/1-16   41.922Mi ± 24%   9.481Mi ± 111%  -77.38% (p=0.000 n=10)
Features/2-16    42.10Mi ± 28%   12.14Mi ±  96%  -71.16% (p=0.000 n=10)
Features/3-16   42.403Mi ± 22%   9.934Mi ± 106%  -76.57% (p=0.000 n=10)
Features/4-16    42.39Mi ± 26%   12.97Mi ±  86%  -69.40% (p=0.000 n=10)
geomean          42.20Mi         11.04Mi         -73.85%

              │ bench_run1.txt │           bench_run6.txt            │
              │   allocs/op    │  allocs/op   vs base                │
Features/1-16      681.8k ± 0%   243.3k ± 0%  -64.32% (p=0.000 n=10)
Features/2-16      684.5k ± 0%   245.6k ± 0%  -64.12% (p=0.000 n=10)
Features/3-16      692.1k ± 0%   253.3k ± 0%  -63.40% (p=0.000 n=10)
Features/4-16      694.2k ± 0%   255.5k ± 0%  -63.20% (p=0.000 n=10)
geomean            688.2k        249.4k       -63.76%
```